// src/application/domain/utils/food.types.ts
export enum FoodTypes {
    FISH_CHICKEN_MEAT = 'fish_chicken_meat',
    SODA = 'soda',
    SALAD_VEGETABLE = 'salad_vegetable',
    FRIED_SALT_FOOD = 'fried_salt_food',
    MILK = 'milk',
    BEAN = 'bean',
    FRUITS = 'fruits',
    CANDY_SUGAR_COOKIE = 'candy_suggar_cookie',
    BURGER_SAUSAGE = 'burger_sausage'
}